# deepfence_server_client/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: community@deepfence.io
    name: Deepfence Support
    url: https://deepfence.io
  description: "Deepfence Runtime API provides programmatic control over Deepfence\
    \ microservice securing your container, kubernetes and cloud deployments. The\
    \ API abstracts away underlying infrastructure details like cloud provider, \n\
    container distros, container orchestrator and type of deployment. This is one\
    \ uniform API to manage and control security alerts, policies and response to\
    \ alerts for microservices running anywhere i.e. managed pure greenfield container\
    \ deployments or a mix of containers, VMs and serverless paradigms like AWS Fargate."
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  termsOfService: /tos
  title: Deepfence ThreatMapper
  version: 2.0.0
...
servers:
- url: /
...
paths:
...
  /deepfence/scan/start/compliance:
    post:
      description: Start Compliance Scan on agent or registry
      operationId: startComplianceScan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelScanTriggerReq'
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelScanTriggerResp'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsBadRequestResponse'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Internal Server Error
      security:
      - bearer_token: []
      summary: Start Compliance Scan
      tags:
      - Compliance
  /deepfence/scan/start/malware:
    post:
      description: Start Malware Scan on agent or registry
      operationId: startMalwareScan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelScanTriggerReq'
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelScanTriggerResp'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsBadRequestResponse'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Internal Server Error
      security:
      - bearer_token: []
      summary: Start Malware Scan
      tags:
      - Malware Scan
  /deepfence/scan/start/secret:
    post:
      description: Start Secret Scan on agent or registry
      operationId: startSecretScan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelScanTriggerReq'
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelScanTriggerResp'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsBadRequestResponse'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Internal Server Error
      security:
      - bearer_token: []
      summary: Start Secret Scan
      tags:
      - Secret Scan
  /deepfence/scan/start/vulnerability:
    post:
      description: Start Vulnerability Scan on agent or registry
      operationId: startVulnerabilityScan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelScanTriggerReq'
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelScanTriggerResp'
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsBadRequestResponse'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiDocsFailureResponse'
          description: Internal Server Error
      security:
      - bearer_token: []
      summary: Start Vulnerability Scan
      tags:
      - Vulnerability
...
components:
  schemas:
    ApiDocsBadRequestResponse:
      properties:
        error_fields:
          additionalProperties:
            type: string
          nullable: true
          type: object
        message:
          type: string
        success:
          example: false
          type: boolean
      type: object
    ApiDocsFailureResponse:
      properties:
        message:
          type: string
        success:
          example: false
          type: boolean
      type: object
...
    ModelScanTriggerReq:
      example:
        resource_type: image
        resource_id: resource_id
        node_id: node_id
      properties:
        node_id:
          type: string
        resource_id:
          type: string
        resource_type:
          enum:
          - image
          - host
          - container
          type: string
      required:
      - node_id
      - resource_id
      - resource_type
      type: object
    ModelScanTriggerResp:
      example:
        scan_id: scan_id
      properties:
        scan_id:
          type: string
      required:
      - scan_id
      type: object
...
  securitySchemes:
    bearer_token:
      bearerFormat: JWT
      description: RW Access
      scheme: bearer
      type: http